Extract recipe run workflow evidence helpers
